requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
xgboost>=2.0.0
scikit-learn>=1.3.0
optuna>=3.4.0

# Data & Storage
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
chromadb>=0.4.0
sqlalchemy>=2.0.0

# Evaluation & Analysis
shap>=0.43.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# SLM (Phase 5)
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
